+++ scripts/coverage.sh
@@ -1,0 +1,2 @@
+mkdir -p outputs
+pytest --cov-report html:outputs/_covhtml --cov=py2max tests