+++ tests/test_coll.py
@@ -1,0 +1,18 @@
+
+from py2max import Patcher
+
+
+def test_coll():
+    p = Patcher('outputs/test_coll.maxpat')
+    p.add_coll('store', dictionary={
+        1: [1, 1.5, "sam"],
+        2: [1.5, 2, "ok"],
+        3: ["A", 25, 1.2],
+        'a': ['abcd'],
+        'b': [1,2,3,4,5,4],
+    })
+    p.save()
+
+
+if __name__ == '__main__':
+    test_coll()